Removing update layout
